+++ Assets/Scripts/Games/PajamaParty/PajamaParty.cs
@@ -75,10 +75,10 @@ namespace HeavenStudio.Games
         CtrPillowMonkey[,] monkeys;
 
         //cues while unoaded
-        static float WantThreeJump = Single.MinValue;
-        static float WantFiveJump = Single.MinValue;
-        static float WantThrowSequence = Single.MinValue;
-        static float WantSleepSequence = Single.MinValue;
+        static double WantThreeJump = double.MinValue;
+        static double WantFiveJump = double.MinValue;
+        static double WantThrowSequence = double.MinValue;
+        static double WantSleepSequence = double.MinValue;
         static bool WantSleepType = false;
         static int WantSleepAction = (int) PajamaParty.SleepType.Normal;
 
@@ -129,31 +129,31 @@ namespace HeavenStudio.Games
             }
         }
 
-        public override void OnGameSwitch(float beat)
+        public override void OnGameSwitch(double beat)
         {
-            if (WantThreeJump != Single.MinValue)
+            if (WantThreeJump != double.MinValue)
             {
                 DoThreeJump(WantThreeJump, false);
-                WantThreeJump = Single.MinValue;
+                WantThreeJump = double.MinValue;
             }
-            if (WantFiveJump != Single.MinValue)
+            if (WantFiveJump != double.MinValue)
             {
                 DoFiveJump(WantFiveJump, false);
-                WantFiveJump = Single.MinValue;
+                WantFiveJump = double.MinValue;
             }
-            if (WantThrowSequence != Single.MinValue)
+            if (WantThrowSequence != double.MinValue)
             {
                 DoThrowSequence(WantThrowSequence, false);
-                WantThrowSequence = Single.MinValue;
+                WantThrowSequence = double.MinValue;
             }
-            if (WantSleepSequence != Single.MinValue)
+            if (WantSleepSequence != double.MinValue)
             {
                 DoSleepSequence(WantSleepSequence, WantSleepType, WantSleepAction, false);
-                WantSleepSequence = Single.MinValue;
+                WantSleepSequence = double.MinValue;
             }
         }
 
-        public void DoThreeJump(float beat, bool doSound = true)
+        public void DoThreeJump(double beat, bool doSound = true)
         {
             Mako.ScheduleJump(beat);
             if (doSound)
@@ -188,7 +188,7 @@ namespace HeavenStudio.Games
             });
         }
 
-        public static void WarnThreeJump(float beat)
+        public static void WarnThreeJump(double beat)
         {
             MultiSound.Play(new MultiSound.Sound[] { 
                 new MultiSound.Sound("pajamaParty/three1", beat), 
@@ -198,7 +198,7 @@ namespace HeavenStudio.Games
             WantThreeJump = beat;
         }
 
-        public void DoFiveJump(float beat, bool doSound = true)
+        public void DoFiveJump(double beat, bool doSound = true)
         {
             Mako.ScheduleJump(beat);
             if (doSound)
@@ -220,7 +220,7 @@ namespace HeavenStudio.Games
             });
         }
 
-        public static void WarnFiveJump(float beat)
+        public static void WarnFiveJump(double beat)
         {
             MultiSound.Play(new MultiSound.Sound[] { 
                 new MultiSound.Sound("pajamaParty/five1", beat), 
@@ -232,7 +232,7 @@ namespace HeavenStudio.Games
             WantFiveJump = beat;
         }
 
-        public void DoThrowSequence(float beat, bool doSound = true)
+        public void DoThrowSequence(double beat, bool doSound = true)
         {
             Mako.ScheduleThrow(beat);
             if (doSound)
@@ -245,13 +245,13 @@ namespace HeavenStudio.Games
             });
         }
 
-        public static void WarnThrowSequence(float beat)
+        public static void WarnThrowSequence(double beat)
         {
             PlayThrowSequenceSound(beat, true);
             WantThrowSequence = beat;
         }
 
-        public static void PlayThrowSequenceSound(float beat, bool force = false)
+        public static void PlayThrowSequenceSound(double beat, bool force = false)
         {
             MultiSound.Play(new MultiSound.Sound[] { 
                 new MultiSound.Sound("pajamaParty/throw1", beat), 
@@ -265,7 +265,7 @@ namespace HeavenStudio.Games
             }, forcePlay: force);
         }
 
-        public void DoSleepSequence(float beat, bool alt = false, int action = (int) PajamaParty.SleepType.Normal, bool doSound = true)
+        public void DoSleepSequence(double beat, bool alt = false, int action = (int) PajamaParty.SleepType.Normal, bool doSound = true)
         {
             var cond = Conductor.instance;
             Mako.StartSleepSequence(beat, alt, action);
@@ -280,7 +280,7 @@ namespace HeavenStudio.Games
                 });
         }
 
-        public static void WarnSleepSequence(float beat, bool alt = false, int action = (int) PajamaParty.SleepType.Normal)
+        public static void WarnSleepSequence(double beat, bool alt = false, int action = (int) PajamaParty.SleepType.Normal)
         {
             MultiSound.Play(new MultiSound.Sound[] { 
                 new MultiSound.Sound("pajamaParty/siesta1", beat), 
@@ -299,7 +299,7 @@ namespace HeavenStudio.Games
             Bed.GetComponent<Animator>().Play("BedImpact", -1, 0);
         }
 
-        public void JumpRow(int row, float beat, int alt = 1)
+        public void JumpRow(int row, double beat, int alt = 1)
         {
             if (row > 4 || row < 0)
             {
@@ -314,7 +314,7 @@ namespace HeavenStudio.Games
             }
         }
 
-        public void JumpCol(int col, float beat, int alt = 1)
+        public void JumpCol(int col, double beat, int alt = 1)
         {
             if (col > 4 || col < 0)
             {
@@ -329,7 +329,7 @@ namespace HeavenStudio.Games
             }
         }
 
-        public void MonkeyCharge(float beat)
+        public void MonkeyCharge(double beat)
         {
             foreach (CtrPillowMonkey monkey in monkeys)
             {
@@ -340,7 +340,7 @@ namespace HeavenStudio.Games
             }
         }
 
-        public void MonkeyThrow(float beat)
+        public void MonkeyThrow(double beat)
         {
             foreach (CtrPillowMonkey monkey in monkeys)
             {
@@ -351,7 +351,7 @@ namespace HeavenStudio.Games
             }
         }
 
-        public void MonkeySleep(float beat, int action)
+        public void MonkeySleep(double beat, int action)
         {
             foreach (CtrPillowMonkey monkey in monkeys)
             {
